Wrap net.Dialer.Dial when there's no custom non-TLS dialer specified
The AIA transport we're using only defines DialTLS. This means we
weren't specifying a Dial/DialContext at all, so HTTP requests to IPs in
denyHosts weren't being restricted.

With this commit, we ensure we always specify Dial or DialContext. If the
transport wrapped specify one of the two, we wrap that one. If it
doesn't specify any of them, we specify DialContext, since Dial is
deprecated, and in this way we ensure covering all cases.

We don't need to do the same for DialTLS since if there's no custom
DialTLS specified, the transport just use dial (and any custom
Dial/DialContext) combined with the TLS handshake.